Week 9/app/requirements.txt
transformers==4.1.1 
google-cloud-storage==1.35.0 
scikit-learn==0.24.0 
pandas==1.1.5
streamlit==1.12.0
google-cloud-aiplatform==1.18.2